Updated Websocket API to add Candles Channel. Updated timestamp on Candles to be i64 (Unix Time, time_t). README updates.

+++ src/product.rs
@@ -218,7 +218,7 @@ pub struct ProductBook {
 pub struct Candle {
     /// Timestamp for bucket start time, in UNIX time.
     #[serde_as(as = "DisplayFromStr")]
-    pub start: f64,
+    pub start: i64,
     /// Lowest price during the bucket interval.
     #[serde_as(as = "DisplayFromStr")]
     pub low: f64,
